backend/requirements.txt
# MedVerify Backend Dependencies (Version Pinned)

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
aiofiles==23.2.1

# --- Data Validation ---
pydantic==2.7.0

# --- Deep Learning ---
torch==2.3.0
torchvision==0.18.0

# --- OCR ---
easyocr==1.7.1
pytesseract==0.3.10
paddleocr==2.8.1
paddlepaddle==3.0.0

# --- Computer Vision ---
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# --- Database ---
sqlalchemy==2.0.30

# --- HTTP client ---
httpx==0.27.0
google-generativeai==0.8.5

# --- Auth / Rate limiting / Config ---
python-jose==3.3.0
slowapi==0.1.9
python-dotenv==1.0.1

# --- ML Training & Data Processing ---
scikit-learn==1.4.0
matplotlib==3.8.0
pandas==2.2.2
openpyxl==3.1.2
kaggle==1.6.14
kagglehub==0.3.13

# --- LLM (local inference fallback) ---
llama-cpp-python==0.2.77
huggingface-hub==0.23.0

# --- Testing ---
pytest==8.2.0
pytest-asyncio==0.23.6
